fix(managed-agents): use shared log viewer
Summary:
- Move the generic log viewer components into shared with compatibility
  re-exports at their previous paths.
- Replace Managed Agents log action terminal icons with the standard eye
  icon and route agent/session logs through the shared viewer.
- Add managed-agent log adapters, Storybook coverage, snapshots, and
  focused tests for the icon/viewer behavior.

Rationale:
- Borg UI uses the eye affordance for viewing logs elsewhere, so Managed
  Agents should match that convention.
- Keeping the log viewer in shared avoids another page-local log dialog
  and makes future reuse explicit.

Tests:
- cd frontend && npm run check:locales
- cd frontend && npm run typecheck
- cd frontend && npm run lint
- cd frontend && npm run build
- cd frontend && npm run test -- src/components/__tests__/LogViewerDialog.test.tsx src/components/__tests__/TerminalLogViewer.test.tsx --run
- cd frontend && npm run test -- src/pages/__tests__/ManagedAgents.test.tsx --run
- cd frontend && npm run snapshots
- Storybook browser walkthrough for Managed Agents log icons and viewers
- git diff --check
